refactor: rename colour by color
